MilicaCosminSergiu_licenta/MilicaCosminSergiu_licenta/PlatformaWeb/requirements.txt
fastapi~=0.104.1
uvicorn
peewee~=3.17.0
pymysql
requests
requests~=2.31.0
grpcio
PyJWT
protobuf
pydantic
